Exceptions

+++ src/main/java/com/zukeramrenzo/course/services/UserService.java
@@ -2,7 +2,11 @@ package com.zukeramrenzo.course.services;
 
 import com.zukeramrenzo.course.entities.User;
 import com.zukeramrenzo.course.repositories.UserRepository;
+import com.zukeramrenzo.course.services.exceptions.DatabaseException;
+import com.zukeramrenzo.course.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,7 +23,7 @@ public class UserService {
 
     public User findById(Long id){
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insert(User obj){
@@ -27,13 +31,26 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try{
+            // The old method that catch the EmptyResultDataAccessException doesn't work anymore
+            if(repository.existsById(id)){
+                repository.deleteById(id);
+            }else{
+                throw new ResourceNotFoundException(id);
+            }
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
-        User entity = repository.getReferenceById(id);
-        updateData(entity, obj);
-        return repository.save(entity);
+        try{
+            User entity = repository.getReferenceById(id);
+            updateData(entity, obj);
+            return repository.save(entity);
+        }catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User obj) {
